Conidtion only on top note of melody
This commit only changes two lines in the lc audio reader so that
instead of all currently playing notes being added to the lc embeddings,
only the highest note is added. this is usually the melody, so that the
model only sees the embeddings for the melody. hope is that this will
help in cases where many baseline notes are playing like for piano